requirements.txt
annotated-types==0.7.0
anyio==4.4.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mypy-extensions==1.0.0
openai==1.33.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.7.4
pydantic_core==2.18.4
pyTelegramBotAPI==4.19.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
